Checkout: fix action for succesful payment
The action woocommerce_checkout_order_processed is not the correct indicator
if payment was succesful but only if all data are correct.

You can stop the mondu payment process in the popup go back one and
retry it and now there is no payment popup but the order is flagged
as payed while mondu status is cancelled.

woocommerce_payment_successful_result is the correct action.

Signed-Off-By: Sven Auhagen <sven.auhagen@voleatech.de>