Order Pay Page: Support Order Pay page
The order pay page is used when an order is created in the backend
of Woocommerce and the user pays the precreated order.
It is a simplified checkout page as all data are already fixed and
added to the order.
The information are obtained by the url and order id.

Signed-Off-By: Sven Auhagen <sven.auhagen@voleatech.de>